% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateFSPInstance}
\alias{generateFSPInstance}
\title{Generate flowshop instances with given attributes}
\usage{
generateFSPInstance(no_jobs, no_machines, ..., seed = NA)
}
\arguments{
\item{no_jobs}{Number of jobs}

\item{no_machines}{Number of machines}

\item{...}{Parameters passed to Instance(...)}

\item{seed}{Optional random number generator seed}
}
\value{

}
\description{
Generate flowshop instances with given attributes
}
