% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance.R
\docType{class}
\name{Instance-class}
\alias{Instance-class}
\title{Scheduling instance class}
\value{

}
\description{
Scheduling instance class
}
\section{Slots}{

\describe{
\item{\code{.Data}}{unrelated, from matrix class}

\item{\code{distribution_type}}{processing times distribution type (uniform, exponential, binomial or erlang)}

\item{\code{correlation_type}}{processing times correlation type (random, machine-correlated, job-correlated)}

\item{\code{correlation}}{correlation value (between 0 and 1)}
}}

