% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance.R
\name{Instance}
\alias{Instance}
\title{Scheduling instance constructor}
\usage{
Instance(
  data = NA,
  nrow = 1,
  ncol = 1,
  byrow = FALSE,
  dimnames = NULL,
  distribution_type = NA_character_,
  correlation_type = NA_character_,
  correlation = 0,
  seed = NA_integer_
)
}
\arguments{
\item{data}{processing time data}

\item{nrow}{number of jobs}

\item{ncol}{number of machines}

\item{byrow}{processing times order by machine (default) or by job}

\item{dimnames}{name of the processing times matrix dimentions}

\item{distribution_type}{processing times distribution type (uniform, exponential, binomial or erlang)}

\item{correlation_type}{processing times correlation type (random, machine-correlated, job-correlated)}

\item{correlation}{correlation value (between 0 and 1)}

\item{seed}{RNG seed used to generate the processing times}
}
\value{
a instance with given processing times
}
\description{
Scheduling instance constructor
}
\examples{
Instance(1:10, 5, 2, distribution_type = "uniform")
}
