% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_statistics.R
\name{instance_statistics}
\alias{instance_statistics}
\title{Instance processing times statistics}
\usage{
instance_statistics(inst)
}
\arguments{
\item{inst}{an instance}
}
\value{
A list with different statistics of processing times:
\itemize{
\item standard deviation (pt_sd)
\item mean standard deviation per machine (mean_sd_per_machine)
\item mean standard deviation per job (mean_sd_per_job)
\item mean skewness per machine (mean_skew_per_machine)
\item mean skewness per job (mean_skew_per_job)
\item mean kurtosis per machine (mean_kurt_per_machine)
\item mean kurtosis per job (mean_kurt_per_job)
}
}
\description{
Instance processing times statistics
}
\examples{
instance_statistics(generate_fsp_instance(20, 5, "erlang", "machine-correlated", 0.95))
}
