% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance.R
\name{plot,Instance-method}
\alias{plot,Instance-method}
\title{Plot instance as pairs with correlation.}
\usage{
\S4method{plot}{Instance}(x, y, ...)
}
\arguments{
\item{Instance}{instance to be plotted.}
}
\description{
Plot instance as pairs with correlation.
}
\examples{
plot(generateFSPInstance(100, 3))
}
