% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_txt.R
\name{read_txt}
\alias{read_txt}
\title{Reads an instance from a txt file with format:
NumJobs NumMachines
ProcTime11 ProcTime12 ...
ProcTime21 ProcTime22 ...
...}
\usage{
read_txt(file = "data", sep = " ", ...)
}
\arguments{
\item{file}{file connection or path}

\item{sep}{separator used to write dimentions and processing times}

\item{...}{parameters passed to Instance}
}
\value{
Instance object
}
\description{
Reads an instance from a txt file with format:
NumJobs NumMachines
ProcTime11 ProcTime12 ...
ProcTime21 ProcTime22 ...
...
}
\examples{
write_txt(generate_fsp_instance(10, 5), "example")
read_txt("example")
}
