% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_txt.R
\name{write_txt,Instance-method}
\alias{write_txt,Instance-method}
\title{Write instance to a txt file with format:
NumJobs NumMachines
ProcTime11 ProcTime12 ...
ProcTime21 ProcTime22 ...
...}
\usage{
\S4method{write_txt}{Instance}(object, file = "data", append = FALSE, sep = " ")
}
\arguments{
\item{object}{instance to written to a file}

\item{file}{file connection or path}

\item{append}{append instance data to existing file}

\item{sep}{separator used to write dimentions and processing times}
}
\description{
Write instance to a txt file with format:
NumJobs NumMachines
ProcTime11 ProcTime12 ...
ProcTime21 ProcTime22 ...
...
}
\examples{
write_txt(generate_fsp_instance(20, 10))
}
