% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance.R
\docType{class}
\name{Instance-class}
\alias{Instance-class}
\title{Scheduling instance class}
\value{

}
\description{
Scheduling instance class
}
\section{Slots}{

\describe{
\item{\code{distribution_type}}{integer.}

\item{\code{correlation_type}}{integer.}

\item{\code{correlation}}{numeric.}

\item{\code{seed}}{integer.}
}}

