% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance.R
\name{Instance}
\alias{Instance}
\title{Scheduling instance constructor}
\usage{
Instance(
  distribution_type = DISTRIBUTION_TYPES$UNIFORM,
  correlation_type = CORRELATION_TYPES$JOB_CORRELATED,
  correlation = NA_real_,
  ...
)
}
\value{

}
\description{
Scheduling instance constructor
}
\section{Slots}{

\describe{
\item{\code{distribution_type}}{integer.}

\item{\code{correlation_type}}{integer.}

\item{\code{correlation}}{numeric.}

\item{\code{seed}}{integer.}
}}

